Initial version of the `narrowable` proc macro.
This provides a way of turning fat pointers to trait objects into thin pointers.
All one has to do is add the following annotation to a trait:

  #[narrowable(N)]
  trait T { ... }

where `N` is the name of a struct that will be created for thin pointers to
objects of type `T`. Conventionally for a trait `T` we call the struct
`NarrowT`:

  #[narrowable(NarrowT)]
  trait T { ... }

One can then do things like:

  struct S { ... }
  impl T for S { ... }

  let thinptr = NarrowT::new(S{ ... });
  let fatptr = *thinptr;

and so on.

There are two chief difficulties to the implementation. First, it's a proc
macro, and this is not the most elegant part of Rust's design: any errors in the
proc macro are very hard to debug. Second, we have to deal with exotic
bigger-than-word alignments: the details of this are explained in an extensive
comment inside the `struct #struct_id`.