Add a natrob proc macro for abgc.
Unfortunately this is surprisingly dissimilar to the "normal" natrob. The most
fundamental problem is that Gc::alloc_blank gives us a pointer which we have to
return intact to Gc::from_raw. That means we can't use the trick "normal"
natrob uses of returning a pointer *after* the vtable (i.e. an interior
pointer): we have to fall back to our old plan of the layout being `(vtable,
object)`, which means that `object` can't have a bigger alignment than a
`usize`. This isn't likely to cause us many problems in practise, though it is
unsatisfying.

It's also a bit annoying that all proc macros have to reside in lib.rs, but
that's forced upon us by current limitations in rustc.