Make the abgc proc macro generate a sized type.
Zero-sized types in Rust are handled differently than non-zero-sized types: some
of the differences are fairly obvious, but some aren't obvious (to me, at least)
in the face of unsafe code. Since we can quite sensibly make this a sized-type
(as the non-abgc proc macro has always done) it makes sense to do so.