Add a recover_gc function.
This is very similar to the `recover` function although slightly harder (but
definitely not impossible) to abuse. For example if you have a narrow trait type
Obj and a concrete type Class which implements that trait, this code leads to
undefined behaviour:

  let x = Gc::new(Class { ... });
  let to = unsafe { Obj::recover_gc(x) };

because `x` was never a narrow trait object in the first place.