Make sure we can handle arbitrary aligned objects.
Previously we could not handle objects which had an alignment greater than a
usize. This commit changes the way we store things but does so in a way that
doesn't make the expected standard case worse (indeed, the machine code for
yksom remains the same before/after this change).

In essence we store a vtable pointer followed (without padding) by an object.
The vtable pointer is aligned to a machine word; the object will be aligned to
`max(align(usize), align(obj))`. That means that the vtable pointer might not be
at the beginning of the memory block.

For a word (or smaller) aligned object on a 64-bit machine the layout will look
as follows (in bytes):

  0..7: vtable ptr
  8..:  object

i.e. the same as we had before. However, for a 32-byte aligned object the layout
will be:

  0..23:  padding
  24..31: vtable ptr
  32..:   object

In this later case the pointer that we store in the struct is an interior
pointer (pointing to the beginning of the vtable pointer). Since, in general, we
don't expect many people to store objects whose alignment is bigger than a word,
the inevitable slow-down in such a case seems a price worth paying. The base
pointer of the memory block can be recovered using `object`'s alignment
(accessible via its vtable) if necessary.