[narrowable_rboehm]: Don't finalize `ThinObj` if it's unnecessary
The `Drop` method for `ThinObj` acts as simple drop glue to call the
inner `Obj`'s drop method. Unfortunately, this creates false positives
in our `needs_finalize` analysis since rboehm thinks it's always
required.

This change will unregister the finalizer for the narrow trait object if
it's clear that it's unnecessary for the inner type.