Use type punning to expose the instantiated type `T` to the GC.
Previously natrob used Gc::new_from_layout to perform its tricks. However, that
meant that we had to use `Gc::<T>::new_from_layout` to tell the GC the "base
type". This wasn't safe as we didn't know for sure how/where the GC might lay
out its memory (e.g. due to alignment).

This commit changes things so that `natrob::new` now uses type punning to avoid
this problem. In essence we add a new struct and a union so that (when expanded)
natrob looks roughly like:

```
union {
  short: struct T { vtable } // #struct_short_id
  long: struct U { vtable, obj } // #struct_long_id
}
```

The union and the two structs are all `repr(C)` to guarantee layout (C99
guarantees that if a union's constituent structs share the same values then one
can safely pun between them). We also rely on Rust's guarantee that
`ManuallyDrop` does not affect layout. We can then safely move between the
"short" and "long" representations. The nice thing is that this makes clear to
`Gc` what the "base" type is while still keeping us in the land of defined
behaviour.

This assumes that the `Gc<T>` is fine with being type-punned to `Gc<U>` iff `T`
and `U` can be safely type punned. We have implicitly made that guarantee and I
think we can now be clear about why it's useful!

There is a trade-off here: we can no longer store things with large aligments
(in practise we require that `align_of::<#struct_long_id<U>>() ==
align_of::<#struct_long_id<u8>>()`, which is enforced by an `assert`). This
isn't a problem for any of the use cases we have and it *could* be solved by
dynamically reading the trait object's alignment out of its vtable -- but we
don't want to pay that cost.

natrob::new_from_layout still uses the "old" not-strictly-safe layout tricks and
I think the next stage will be to get rid of that entirely (which means changing
yksom).